Revert "split out duties scheduler into separate schedulers (#2636)"
This reverts commit 559c1d7b339611bc19ff3905f92d9a411d4ce9c2.
